Move opb to transform
